+++ src/dolphin/mod.rs
@@ -15,38 +15,51 @@ use process_memory::DolphinProcessMemory;
 pub trait Dolphin {
   /// # Safety
   /// `maddr + size` must be in bound
-  unsafe fn operate_memory_unchecked<T, F>(
+  unsafe fn read_memory_unchecked<T, F>(
     &self, maddr: DolphinMemAddr, size: usize, operator: F,
   ) -> Option<T>
-    where F: FnOnce(*mut u8) -> T;
+    where F: FnOnce(*const u8) -> T;
+
+  /// # Safety
+  /// `maddr + payload.len()` must be in bound
+  unsafe fn write_memory_unchecked(&self, maddr: DolphinMemAddr, payload: &[u8]) -> Option<()>;
 
   #[inline]
-  fn operate_memory<T, F>(&self, addr: Addr, size: usize, operator: F) -> Option<T>
-    where F: FnOnce(*mut u8) -> T
+  fn read_memory<T, F>(&self, addr: Addr, size: usize, operator: F) -> Option<T>
+    where F: FnOnce(*const u8) -> T
   {
     DolphinMemAddr::try_from(addr).ok().and_then(|maddr| {
       if (maddr.space() as usize) < size {return None}
-      unsafe {self.operate_memory_unchecked(maddr, size, operator)}
+      unsafe {self.read_memory_unchecked(maddr, size, operator)}
     })
   }
 
   #[inline]
-  fn operate_memory_truncated<T, F>(&self, addr: Addr, max_size: usize, operator: F) -> Option<T>
-    where F: FnOnce(*mut u8, usize) -> T
+  fn read_memory_truncated<T, F>(&self, addr: Addr, max_size: usize, operator: F) -> Option<T>
+    where F: FnOnce(*const u8, usize) -> T
   {
     DolphinMemAddr::try_from(addr).ok().and_then(|maddr| {
       let size = std::cmp::min(maddr.space() as usize, max_size);
-      unsafe {self.operate_memory_unchecked(maddr, size, |ptr| operator(ptr, size))}
+      unsafe {self.read_memory_unchecked(maddr, size, |ptr| operator(ptr, size))}
+    })
+  }
+
+  #[inline]
+  fn write_bytes(&self, addr: Addr, payload: &[u8]) -> Option<()> {
+    let size = payload.len();
+    DolphinMemAddr::try_from(addr).ok().and_then(|maddr| {
+      if (maddr.space() as usize) < size {return None}
+      unsafe {self.write_memory_unchecked(maddr, payload)}
     })
   }
 
   fn read<T: DecodeBE>(&self, addr: Addr) -> Option<T> {
     let size = std::mem::size_of::<T>();
-    self.operate_memory(addr, size, |ptr| unsafe {T::decode_be(ptr)})
+    self.read_memory(addr, size, |ptr| unsafe {T::decode_be(ptr)})
   }
   fn read_str(&self, addr: Addr) -> Option<String> {
     let maxlen = 256; // TODO
-    self.operate_memory_truncated(addr, maxlen, |ptr, maxlen| {
+    self.read_memory_truncated(addr, maxlen, |ptr, maxlen| {
       let mut len = 0usize;
       while len < maxlen && unsafe{*ptr.add(len)} != 0 {
         len += 1;
@@ -57,17 +70,10 @@ pub trait Dolphin {
     }).unwrap_or(None)
   }
   fn dump_hex(&self, addr: Addr, size: usize) -> Option<String> {
-    self.operate_memory(addr, size, |ptr| {
+    self.read_memory(addr, size, |ptr| {
       (0..size)
         .map(|i| format!("{:02X}", unsafe {*ptr.add(i)}))
         .collect()
-    })
-  }
-
-  fn write_bytes(&self, addr: Addr, payload: &[u8]) -> Option<()> {
-    let size = payload.len();
-    self.operate_memory(addr, size, |ptr| unsafe {
-      std::ptr::copy(payload.as_ptr(), ptr, size);
     })
   }
 }
@@ -77,12 +83,18 @@ pub enum DolphinMemory {
   ProcessMemory(DolphinProcessMemory),
 }
 impl Dolphin for DolphinMemory {
-  unsafe fn operate_memory_unchecked<T, F>(&self, maddr: DolphinMemAddr, size: usize, operator: F) -> Option<T>
-    where F: FnOnce(*mut u8) -> T
+  unsafe fn read_memory_unchecked<T, F>(&self, maddr: DolphinMemAddr, size: usize, operator: F) -> Option<T>
+    where F: FnOnce(*const u8) -> T
   {
     match self {
-      DolphinMemory::SharedMemory(m) => m.operate_memory_unchecked(maddr, size, operator),
-      DolphinMemory::ProcessMemory(m) => m.operate_memory_unchecked(maddr, size, operator),
+      DolphinMemory::SharedMemory(m) => m.read_memory_unchecked(maddr, size, operator),
+      DolphinMemory::ProcessMemory(m) => m.read_memory_unchecked(maddr, size, operator),
+    }
+  }
+  unsafe fn write_memory_unchecked(&self, maddr: DolphinMemAddr, payload: &[u8]) -> Option<()> {
+    match self {
+      DolphinMemory::SharedMemory(m) => m.write_memory_unchecked(maddr, payload),
+      DolphinMemory::ProcessMemory(m) => m.write_memory_unchecked(maddr, payload),
     }
   }
 }